requirements.txt
scrapy 
scrapy-playwright 
playwright
pydispatcher
beautifulsoup4
lxml